pyproject.toml
[project]
name = "cdhdb"
version = "0.1.0"
description = "A template for data harvesting, processing, and serving solutions"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.11.13",
    "asyncio>=3.4.3",
    "jsonschema>=4.23.0",
    "lxml>=5.3.1",
    "pandas>=2.2.3",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.38",
    "tqdm>=4.67.1",
]
